gui-integration: connected buttons and spinbox events, managed back navigation

+++ src/menu/menu.c
@@ -27,6 +27,8 @@
 // --------- Global register start ---------
 uint8_t menu_tot_events; 
 menu_event_t  menu_events[menu_event_max];
+menu_button_group_type menu_nav_stack[MENU_TOTAL_NAV_STACK];
+uint8_t menu_button_index = 0;
 
 static graphic_ctx_t menu_graphic_ctx = {
     .bppx = MENU_OVERLAY_BBPX,
@@ -43,6 +45,13 @@ const gui_properties_t menu_common_nshared_props = {
     .shared     = 0,
     .border     = 1
 };
+const gui_properties_t menu_spinbox_props = {
+		.alignment = gui_align_center,
+		.horiz_vert = gui_orientation_horizontal,
+		.padding = 1,
+		.shared = 0,
+		.border = 1
+	};
 
 gui_object_t *menu_focused_object = NULL;
 gui_object_t menu_window;
@@ -69,7 +78,11 @@ void menu_on_keyboard_event(keyboard_status_t keys) {
                         gui_unfocused(menu_focused_object);
                         menu_focused_object = gui_focused(&menu_left_buttons_group_elements[0]);
                     } else if (menu_focused_object != NULL) {
-                        menu_focused_object = gui_deactivate(menu_focused_object);
+                        gui_object_t *old_focus = menu_focused_object;
+                        gui_object_t *new_focus = gui_deactivate(menu_focused_object);
+                        if (menu_focused_object == old_focus && new_focus != old_focus) {
+                            menu_focused_object = new_focus;
+                        }
                     }
                 } else if (event->key_down && is_video_overlay_enabled() && menu_focused_object != NULL) {
                     menu_focused_object = gui_activate(menu_focused_object);
@@ -123,6 +136,47 @@ bool on_exit_event(gui_status_t status, gui_base_t *origin, gui_object_t *destin
     return true;
 }
 
+bool on_alignment_event(gui_status_t status, gui_base_t *origin, gui_object_t *destination) {
+    if (!status.activated && origin->status.activated) {
+        menu_button_group_type previous = menu_nav_stack[menu_button_index - 1];
+        menu_button_group_type new = menu_button_group_alignment;
+        menu_left_buttons_group = menu_create_left_button_group(previous, new);
+        gui_obj_draw(menu_window);
+        gui_obj_draw(menu_left_buttons_group);
+        menu_focused_object = gui_focused(&menu_left_buttons_group_elements[0]);
+    }
+    return true;
+}
+
+bool on_back_event(gui_status_t status, gui_base_t *origin, gui_object_t *destination) {
+    if (!status.activated && origin->status.activated) {
+        if (menu_button_index >= 2) {
+            menu_button_group_type previous = menu_nav_stack[menu_button_index - 1];
+            menu_button_group_type new = menu_nav_stack[menu_button_index - 2];
+            menu_button_index -= 2;
+            menu_left_buttons_group = menu_create_left_button_group(previous, new);
+            gui_obj_draw(menu_window);
+            gui_obj_draw(menu_left_buttons_group);
+            menu_focused_object = gui_focused(&menu_left_buttons_group_elements[0]);
+        }
+    }
+    return false;
+}
+
+uint test1, test2;
+bool on_spinbox_event(gui_status_t status, gui_base_t *origin, gui_object_t *destination) {
+		if (!status.activated && origin->status.activated) {
+			destination->base.status.navigable = !destination->base.status.navigable;
+		} else if (status.add && test1 < 100) {
+			test1 += 1;
+		} else if (status.substract && test1 != 0) {
+			test1 -= 1;
+		}
+		destination->base.status.data_changed = 1;
+		
+		return true;
+	}
+
 // ----------  GUI EVENT SLOTS HANDLERS END  ---------- 
 
 void menu_elements_copy_(const gui_object_t *src, gui_object_t *dst, uint8_t size) {
@@ -131,10 +185,30 @@ void menu_elements_copy_(const gui_object_t *src, gui_object_t *dst, uint8_t siz
     }
 }
 
-gui_object_t menu_create_left_button_group(menu_button_group_type type) {
-    switch(type) {
+gui_object_t menu_create_left_button_group(menu_button_group_type previous, menu_button_group_type new) {
+    //First unsubscribe last objects
+      
+    switch(previous) {
+        case menu_button_group_home:
+            gui_event_unsubscribe(&menu_left_buttons_group_elements[0].base, &menu_left_buttons_group_elements[0]);
+            gui_event_unsubscribe(&menu_left_buttons_group_elements[3].base, &menu_left_buttons_group_elements[3]);
+            break;
+        case menu_button_group_alignment:
+            gui_event_unsubscribe(&menu_left_buttons_group_elements[2].base, &menu_left_buttons_group_elements[2]);
+            break;
+        default:
+            break;
+    };
+    // Avoid new button groups if stack is full
+    if (menu_button_index > MENU_TOTAL_LEFT_BUTTONS) { 
+        return menu_left_buttons_group; 
+    }
+    menu_button_index++;
+    //Generate new objects and subscriptions
+    menu_nav_stack[menu_button_index - 1] = new;
+    switch(new) {
         case menu_button_group_home: {
-             gui_object_t elements[] = { 
+            gui_object_t elements[] = { 
                 gui_create_button(&menu_overlay_ctx,  0, 0, 120, 12, &menu_colors_list, menu_common_nshared_props, "Alignment"),
                 gui_create_button(&menu_overlay_ctx,  0, 0, 120, 12, &menu_colors_list, menu_common_nshared_props, "Diagnostics"),
                 gui_create_button(&menu_overlay_ctx,  0, 0, 120, 12, &menu_colors_list, menu_common_nshared_props, "Save & reboot"),
@@ -143,9 +217,27 @@ gui_object_t menu_create_left_button_group(menu_button_group_type type) {
             menu_elements_copy(elements, menu_left_buttons_group_elements);
             gui_list_t group_list = initalizeGuiDynList(menu_left_buttons_group_elements, arraySize(elements));
             menu_left_buttons_group_list = group_list;
-            }
+            
             gui_status_t button_status = { .activated = 1 };
             gui_event_subscribe(button_status, &menu_left_buttons_group_elements[3].base, &menu_left_buttons_group_elements[3], on_exit_event);
+            gui_event_subscribe(button_status, &menu_left_buttons_group_elements[0].base, &menu_left_buttons_group_elements[0], on_alignment_event);
+            }
+            break;
+
+        case menu_button_group_alignment: {
+            gui_object_t elements[] = {
+                gui_create_spinbox(&menu_overlay_ctx, 0, 0, 120, 12, &menu_colors_list, menu_spinbox_props, &test1),
+                gui_create_spinbox(&menu_overlay_ctx, 0, 0, 120, 12, &menu_colors_list, menu_spinbox_props, &test2),
+                gui_create_button(&menu_overlay_ctx,  0, 0, 120, 12, &menu_colors_list, menu_common_nshared_props, "Back")
+            };
+            menu_elements_copy(elements, menu_left_buttons_group_elements);
+            gui_list_t group_list = initalizeGuiDynList(menu_left_buttons_group_elements, arraySize(elements));
+            menu_left_buttons_group_list = group_list;
+            
+            gui_status_t button_status = { .activated = 1, .add = 1, .substract = 1};
+            gui_event_subscribe(button_status, &menu_left_buttons_group_elements[2].base, &menu_left_buttons_group_elements[2], on_back_event);
+            gui_event_subscribe(button_status, &menu_left_buttons_group_elements[0].base, &menu_left_buttons_group_elements[0], on_spinbox_event);
+            }
             break;
             
         default: {
@@ -191,7 +283,7 @@ int menu_initialize(uint *pins, menu_event_type *events, uint8_t count) {
 
     //Menu System initialize
     menu_window = gui_create_window(&menu_graphic_ctx, 0, 0, menu_overlay_ctx.width, menu_overlay_ctx.height, &menu_colors_list, menu_common_nshared_props);
-    menu_left_buttons_group = menu_create_left_button_group(menu_button_group_home);
+    menu_left_buttons_group = menu_create_left_button_group(menu_button_group_none, menu_button_group_home);
     menu_left_buttons_group.base.status.enabled = false;
 
     return 0;
